+++ apps/backend/src/health/health.route.ts
@@ -1,0 +1,7 @@
+import type { FastifyInstance, FastifyRequest } from "fastify";
+
+export async function healthRoute(fastify: FastifyInstance) {
+  fastify.get("/api/health", async (_request: FastifyRequest) => {
+    return { status: "ok" };
+  });
+}
