+++ apps/backend/src/health/index.ts
@@ -1,0 +1,1 @@
+export * from "./health.route.js";
